# Speed-up setup_intermediate_state while maintaining trace-ability.