--- a/pytest.ini
+++ b/pytest.ini
@@ -1,7 +1,7 @@
 [pytest]
 xfail_strict = true
 verbosity_assertions = 2
-addopts = -svx --hypothesis-profile=no-deadline
+addopts = -sv --hypothesis-profile=no-deadline
 testpaths =
     tests/
 markers =