--- a/tests/test_examples.py
+++ b/tests/test_examples.py
@@ -259,14 +259,14 @@ def test_run_rbfe_legs(
             "59ca76d1c5d43925ae6b163667c4b3e60f6c765ce10ad85d4f1d14c7d7854933",
         ),
         "solvent": (
-            "b4af69c91a1cfdbdf911c27f929de56f2436b20d5c6d0b3bc3f59021065df2a2",
+            "d6464cc055e486acf6d6d1311aadc90d0ba27118d818be3750d569ace27f39bf",
             "5c0f5fea16e7fad695fe48c4d7e9400de1ee081616e66d81c13cff975275785d",
             "18478e891fe7e5448df89c10e8eedba32f86bdf572a01749c63957da0dc315f6",
         ),
         "complex": (
-            "1fb7f275f66dc0c4122dbdeeed488074c5e20065aa7836116937d98710619479",
-            "4b858b8b851e7dc642b58ead7468e1ddc6aac13ed7b5961768e75d8d638b3f41",
-            "81ec6f3634d110a8b360a97f09a40f562cd5958aff48c646ae4d0a45c22c6582",
+            "708233cf3dacaa9fe62ba7faf5f873bbe1ea34d1a8b8a5cf80a48d9553f65dc3",
+            "a9269276f75cc64dea521ea723a8af1f1ec7aa4dc99f79c6c940db66b3e576c6",
+            "cefc669291e5027c661be7eb9a3db225f980b1a5ab1b1dfa898da21eea2918ad",
         ),
     }
     with resources.as_file(resources.files("timemachine.testsystems.fep_benchmark.hif2a")) as hif2a_dir: