@proteneer 2025-03-18
I also verified on 25 edges, 12 windows each, that the interpolated parameters on master and this branch are within machine precision of each other:

``` python

# compare.py
from jax import config
config.update("jax_enable_x64", True)

from timemachine.utils import path_to_internal_file
import numpy as np
from timemachine.ff import Forcefield
from timemachine.fe import atom_mapping
from timemachine.fe.single_topology import SingleTopology
from timemachine.constants import DEFAULT_ATOM_MAPPING_KWARGS
from timemachine.fe.utils import get_mol_name
import pickle
from rdkit import Chem

def read_sdf(fname, removeHs: bool = False) -> list[Chem.Mol]:
    supplier = Chem.SDMolSupplier(str(fname), removeHs=removeHs)
    mols = [mol for mol in supplier]
    return mols

def _get_core_by_mcs(mol_a, mol_b):
    all_cores = atom_mapping.get_cores(
        mol_a,
        mol_b,
        **DEFAULT_ATOM_MAPPING_KWARGS,
    )

    core = all_cores[0]
    return core

def test_hif2a_end_state_stability(num_pairs_to_setup=25):
    """
    Pick some random pairs from the hif2a set and ensure that they're numerically stable at the
    end-states under a distance based atom-mapping protocol. For a subset of them, we will also run
    simulations.
    """

    seed = 2024

    with path_to_internal_file("timemachine.testsystems.data", "ligands_40.sdf") as path_to_ligand:
        mols = read_sdf(path_to_ligand)

    pairs = [(mol_a, mol_b) for mol_a in mols for mol_b in mols]

    np.random.seed(seed)
    np.random.shuffle(pairs)
    ff = Forcefield.load_from_file("smirnoff_1_1_0_sc.py")
    res = []
    for pair_idx, (mol_a, mol_b) in enumerate(pairs[:num_pairs_to_setup]):
        print("Checking", get_mol_name(mol_a), "->", get_mol_name(mol_b))
        core = _get_core_by_mcs(mol_a, mol_b)
        st = SingleTopology(mol_a, mol_b, core, ff)
        for lamb in np.linspace(0, 1.0, 24):
            gs = st.setup_intermediate_state(lamb)
            res.append((
                gs.bond.params,
                gs.angle.params,
                gs.proper.params,
                gs.improper.params,
                gs.chiral_atom.params,
                gs.nonbonded_pair_list.params))

    with open("res_branch.npz", "wb") as fh:
        pickle.dump(res, fh)

if __name__ == "__main__":
    test_hif2a_end_state_stability(25)
```

``` python
# diff.py
import pickle
import numpy as np

with open("res_branch.npz", "rb") as fh:
    ref_objs = pickle.load(fh)

with open("res_master.npz", "rb") as fh:
    test_objs = pickle.load(fh)

for a,b in zip(ref_objs, test_objs):
    for idx, (aa, bb) in enumerate(zip(a, b)):
        np.testing.assert_allclose(aa, bb)
```