@proteneer 2025-03-18
Thanks @badisa  for helping me track down an annoying subtley:

While this PR emits the exact same parameters for potential objects in 32-bit, it fails to do preserve it in 64-bit. This unfortunately leads to different bitwise results. 

We can try some tricks to improve the behavior of this (eg. by round tripping 64 bit -> 32 bit -> 64bit), so very small tolerances in 64bit are still permitted.